+++ src/VinnyLibConverterCommon/Transformation/QuaternionInfo.cs
@@ -24,7 +24,7 @@ namespace VinnyLibConverterCommon.Transformation
         /// <param name="qy"></param>
         /// <param name="qz"></param>
         /// <param name="qw"></param>
-        public QuaternionInfo(float qx, float qy, float qz, float qw)
+        public QuaternionInfo(double qx, double qy, double qz, double qw)
         {
             this.X = qx;
             this.Y = qy;
@@ -40,25 +40,25 @@ namespace VinnyLibConverterCommon.Transformation
         /// <param name="x"></param>
         /// <param name="y"></param>
         /// <param name="z"></param>
-        public QuaternionInfo (float x = 0f, float y = 0f, float z = 0f)
+        public QuaternionInfo (double x = 0f, double y = 0f, double z = 0f)
         {
-            float xRad = x * 0.5f;
-            float yRad = y * 0.5f;
-            float zRad = z * 0.5f;
+            double xRad = x * 0.5f;
+            double yRad = y * 0.5f;
+            double zRad = z * 0.5f;
 
             // Calculate trigonometric values
-            float sinX = (float)Math.Sin(xRad);
-            float cosX = (float)Math.Cos(xRad);
-            float sinY = (float)Math.Sin(yRad);
-            float cosY = (float)Math.Cos(yRad);
-            float sinZ = (float)Math.Sin(zRad);
-            float cosZ = (float)Math.Cos(zRad);
+            double sinX = (double)Math.Sin(xRad);
+            double cosX = (double)Math.Cos(xRad);
+            double sinY = (double)Math.Sin(yRad);
+            double cosY = (double)Math.Cos(yRad);
+            double sinZ = (double)Math.Sin(zRad);
+            double cosZ = (double)Math.Cos(zRad);
 
             // Calculate quaternion components
-            float qx = sinX * cosY * cosZ - cosX * sinY * sinZ;
-            float qy = cosX * sinY * cosZ + sinX * cosY * sinZ;
-            float qz = cosX * cosY * sinZ - sinX * sinY * cosZ;
-            float qw = cosX * cosY * cosZ + sinX * sinY * sinZ;
+            double qx = sinX * cosY * cosZ - cosX * sinY * sinZ;
+            double qy = cosX * sinY * cosZ + sinX * cosY * sinZ;
+            double qz = cosX * cosY * sinZ - sinX * sinY * cosZ;
+            double qw = cosX * cosY * cosZ + sinX * sinY * sinZ;
 
             this.X = qx;
             this.Y = qy;
@@ -96,14 +96,14 @@ namespace VinnyLibConverterCommon.Transformation
             Vector3 right = Vector3.Cross(up.Normalized, forward).Normalized;
             up = Vector3.Cross(forward, right);
 
-            float num1 = right.X + up.Y + forward.Z;
+            double num1 = right.X + up.Y + forward.Z;
             QuaternionInfo quaternion = new QuaternionInfo();
 
             if (num1 > 0.0f)
             {
-                float num2 = (float)Math.Sqrt(num1 + 1.0f);
+                double num2 = (double)Math.Sqrt(num1 + 1.0f);
                 quaternion.W = num2 * 0.5f;
-                float num3 = 0.5f / num2;
+                double num3 = 0.5f / num2;
                 quaternion.X = (up.Z - forward.Y) * num3;
                 quaternion.Y = (forward.X - right.Z) * num3;
                 quaternion.Z = (right.Y - up.X) * num3;
@@ -112,8 +112,8 @@ namespace VinnyLibConverterCommon.Transformation
 
             if (right.X >= up.Y && right.X >= forward.Z)
             {
-                float num2 = (float)Math.Sqrt(1.0f + right.X - up.Y - forward.Z);
-                float num3 = 0.5f / num2;
+                double num2 = (double)Math.Sqrt(1.0f + right.X - up.Y - forward.Z);
+                double num3 = 0.5f / num2;
                 quaternion.X = 0.5f * num2;
                 quaternion.Y = (right.Y + up.X) * num3;
                 quaternion.Z = (right.Z + forward.X) * num3;
@@ -123,8 +123,8 @@ namespace VinnyLibConverterCommon.Transformation
 
             if (up.Y > forward.Z)
             {
-                float num2 = (float)Math.Sqrt(1.0f + up.Y - right.X - forward.Z);
-                float num3 = 0.5f / num2;
+                double num2 = (double)Math.Sqrt(1.0f + up.Y - right.X - forward.Z);
+                double num3 = 0.5f / num2;
                 quaternion.X = (up.X + right.Y) * num3;
                 quaternion.Y = 0.5f * num2;
                 quaternion.Z = (forward.Y + up.Z) * num3;
@@ -132,8 +132,8 @@ namespace VinnyLibConverterCommon.Transformation
                 return quaternion;
             }
 
-            float num4 = (float)Math.Sqrt(1.0f + forward.Z - right.X - up.Y);
-            float num5 = 0.5f / num4;
+            double num4 = (double)Math.Sqrt(1.0f + forward.Z - right.X - up.Y);
+            double num5 = 0.5f / num4;
             quaternion.X = (forward.X + right.Z) * num5;
             quaternion.Y = (forward.Y + up.Z) * num5;
             quaternion.Z = 0.5f * num4;
@@ -141,17 +141,17 @@ namespace VinnyLibConverterCommon.Transformation
             return quaternion;
         }
 
-        public static QuaternionInfo AngleAxis(float angle, Vector3 axis)
+        public static QuaternionInfo AngleAxis(double angle, Vector3 axis)
         {
             axis = axis.Normalized;
-            float halfAngle = angle * 0.5f * (float)(Math.PI / 180.0);
-            float sin = (float)Math.Sin(halfAngle);
+            double halfAngle = angle * 0.5f * (double)(Math.PI / 180.0);
+            double sin = (double)Math.Sin(halfAngle);
 
             return new QuaternionInfo(
                 axis.X * sin,
                 axis.Y * sin,
                 axis.Z * sin,
-                (float)Math.Cos(halfAngle)
+                (double)Math.Cos(halfAngle)
             );
         }
 
@@ -160,7 +160,7 @@ namespace VinnyLibConverterCommon.Transformation
             fromDirection = fromDirection.Normalized;
             toDirection = toDirection.Normalized;
 
-            float dot = Vector3.Dot(fromDirection, toDirection);
+            double dot = Vector3.Dot(fromDirection, toDirection);
 
             // If vectors are parallel, return identity or 180 degree rotation
             if (dot > 0.99999f)
@@ -176,25 +176,25 @@ namespace VinnyLibConverterCommon.Transformation
             }
 
             Vector3 axisRotation = Vector3.Cross(fromDirection, toDirection).Normalized;
-            float angle = (float)Math.Acos(dot) * (180f / (float)Math.PI);
+            double angle = (double)Math.Acos(dot) * (180f / (double)Math.PI);
 
             return AngleAxis(angle, axisRotation);
         }
 
         public static Vector3 operator *(QuaternionInfo rotation, Vector3 point)
         {
-            float num1 = rotation.X * 2f;
-            float num2 = rotation.Y * 2f;
-            float num3 = rotation.Z * 2f;
-            float num4 = rotation.X * num1;
-            float num5 = rotation.Y * num2;
-            float num6 = rotation.Z * num3;
-            float num7 = rotation.X * num2;
-            float num8 = rotation.X * num3;
-            float num9 = rotation.Y * num3;
-            float num10 = rotation.W * num1;
-            float num11 = rotation.W * num2;
-            float num12 = rotation.W * num3;
+            double num1 = rotation.X * 2f;
+            double num2 = rotation.Y * 2f;
+            double num3 = rotation.Z * 2f;
+            double num4 = rotation.X * num1;
+            double num5 = rotation.Y * num2;
+            double num6 = rotation.Z * num3;
+            double num7 = rotation.X * num2;
+            double num8 = rotation.X * num3;
+            double num9 = rotation.Y * num3;
+            double num10 = rotation.W * num1;
+            double num11 = rotation.W * num2;
+            double num12 = rotation.W * num3;
 
             return new Vector3(
                 (1f - (num5 + num6)) * point.X + (num7 - num12) * point.Y + (num8 + num11) * point.Z,
@@ -205,7 +205,7 @@ namespace VinnyLibConverterCommon.Transformation
 
         public void Normalize()
         {
-            float magnitude = Convert.ToSingle(Math.Sqrt(X * X + Y * Y + Z * Z + W * W));
+            double magnitude = Convert.ToDouble(Math.Sqrt(X * X + Y * Y + Z * Z + W * W));
             if (magnitude > 0)
             {
                 X /= magnitude;
@@ -219,39 +219,39 @@ namespace VinnyLibConverterCommon.Transformation
         /// Возвращает углы Эйлера вдоль осей XYZ из данного определения кватерниона
         /// </summary>
         /// <returns></returns>
-        public float[] GetEulerAnglesRadians()
+        public double[] GetEulerAnglesRadians()
         {
             Normalize();
 
             // Roll (x-axis rotation)
-            float sinr_cosp = 2f * (this.W * this.X + this.Y * this.Z);
-            float cosr_cosp = 1f - 2f * (this.X * this.X + this.Y * this.Y);
-            float roll = (float)Math.Atan2(sinr_cosp, cosr_cosp);
+            double sinr_cosp = 2f * (this.W * this.X + this.Y * this.Z);
+            double cosr_cosp = 1f - 2f * (this.X * this.X + this.Y * this.Y);
+            double roll = (double)Math.Atan2(sinr_cosp, cosr_cosp);
 
             // Pitch (y-axis rotation)
-            float sinp = 2f * (this.W * this.Y - this.Z * this.X);
-            float pitch;
+            double sinp = 2f * (this.W * this.Y - this.Z * this.X);
+            double pitch;
             if (Math.Abs(sinp) >= 1f)
             {
-                if (sinp > 0) pitch = (float)Math.PI / 2f;
-                else pitch = -(float)Math.PI / 2f;
+                if (sinp > 0) pitch = (double)Math.PI / 2f;
+                else pitch = -(double)Math.PI / 2f;
 
-                //pitch = (float)Math.CopySign(Math.PI / 2f, sinp); // Use 90 degrees if out of range
+                //pitch = (double)Math.CopySign(Math.PI / 2f, sinp); // Use 90 degrees if out of range
             }
 
-            else pitch = (float)Math.Asin(sinp);
+            else pitch = (double)Math.Asin(sinp);
 
             // Yaw (z-axis rotation)
-            float siny_cosp = 2f * (this.W * this.Z + this.X * this.Y);
-            float cosy_cosp = 1f - 2f * (this.Y * this.Y + this.Z * this.Z);
-            float yaw = (float)Math.Atan2(siny_cosp, cosy_cosp);
+            double siny_cosp = 2f * (this.W * this.Z + this.X * this.Y);
+            double cosy_cosp = 1f - 2f * (this.Y * this.Y + this.Z * this.Z);
+            double yaw = (double)Math.Atan2(siny_cosp, cosy_cosp);
 
-            return new float[] { roll, pitch, yaw };
+            return new double[] { roll, pitch, yaw };
         }
 
-        public float X { get; set; }
-        public float Y { get; set; }
-        public float Z { get; set; }
-        public float W { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
+        public double Z { get; set; }
+        public double W { get; set; }
     }
 }
