+++ src/VinnyLibConverterCommon/Transformation/Vector3.cs
@@ -6,9 +6,9 @@ namespace VinnyLibConverterCommon.Transformation
 {
     public struct Vector3
     {
-        public float X, Y, Z;
+        public double X, Y, Z;
 
-        public Vector3(float x, float y, float z)
+        public Vector3(double x, double y, double z)
         {
             X = x;
             Y = y;
@@ -24,13 +24,13 @@ namespace VinnyLibConverterCommon.Transformation
         public static readonly Vector3 Right = new Vector3(1, 0, 0);
         public static readonly Vector3 Left = new Vector3(-1, 0, 0);
 
-        public float Magnitude => (float)Math.Sqrt(X * X + Y * Y + Z * Z);
+        public double Magnitude => (double)Math.Sqrt(X * X + Y * Y + Z * Z);
 
         public Vector3 Normalized
         {
             get
             {
-                float mag = Magnitude;
+                double mag = Magnitude;
                 if (mag > 1E-05f)
                     return this / mag;
                 return Zero;
@@ -39,11 +39,11 @@ namespace VinnyLibConverterCommon.Transformation
 
         public static Vector3 operator +(Vector3 a, Vector3 b) => new Vector3(a.X + b.X, a.Y + b.Y, a.Z + b.Z);
         public static Vector3 operator -(Vector3 a, Vector3 b) => new Vector3(a.X - b.X, a.Y - b.Y, a.Z - b.Z);
-        public static Vector3 operator *(Vector3 a, float b) => new Vector3(a.X * b, a.Y * b, a.Z * b);
-        public static Vector3 operator /(Vector3 a, float b) => new Vector3(a.X / b, a.Y / b, a.Z / b);
+        public static Vector3 operator *(Vector3 a, double b) => new Vector3(a.X * b, a.Y * b, a.Z * b);
+        public static Vector3 operator /(Vector3 a, double b) => new Vector3(a.X / b, a.Y / b, a.Z / b);
         public static Vector3 operator -(Vector3 a) => new Vector3(-a.X, -a.Y, -a.Z);
 
-        public static float Dot(Vector3 a, Vector3 b) => a.X * b.X + a.Y * b.Y + a.Z * b.Z;
+        public static double Dot(Vector3 a, Vector3 b) => a.X * b.X + a.Y * b.Y + a.Z * b.Z;
 
         public static Vector3 Cross(Vector3 a, Vector3 b) => new Vector3(
             a.Y * b.Z - a.Z * b.Y,
@@ -51,10 +51,10 @@ namespace VinnyLibConverterCommon.Transformation
             a.X * b.Y - a.Y * b.X
         );
 
-        public static float Distance(Vector3 a, Vector3 b)
+        public static double Distance(Vector3 a, Vector3 b)
         {
             Vector3 vector = new Vector3(a.X - b.X, a.Y - b.Y, a.Z - b.Z);
-            return (float)Math.Sqrt(vector.X * vector.X + vector.Y * vector.Y + vector.Z * vector.Z);
+            return (double)Math.Sqrt(vector.X * vector.X + vector.Y * vector.Y + vector.Z * vector.Z);
         }
 
         public override string ToString() => $"({X:F3}, {Y:F3}, {Z:F3})";
